perf(revisioner): Consolidate schema hydration
