+++ app/revisioner/actions/modified.py
@@ -3,8 +3,8 @@ from app.definitions.models import Schema, Table, Column, Index, IndexColumn
 
 from app.revisioner.models import Revision
 from app.revisioner.collectors import ObjectCollector
-from app.revisioner.revisioners import get_content_type_for_model
 
+from utils.contenttypes import get_content_type_for_model
 from utils.postgres.paginators import RawQuerySetPaginator
 from utils.postgres.types import ConflictAction
 from utils.shortcuts import model_to_dict
